+++ python/aristotle-mdr-api/aristotle_mdr_api/v4/concepts/views.py
@@ -1,0 +1,10 @@
+from rest_framework import generics
+from aristotle_mdr_api.v4.permissions import AuthCanViewEdit
+from aristotle_mdr_api.v4.concepts import serializers
+from aristotle_mdr.models import _concept
+
+
+class ConceptView(generics.RetrieveAPIView):
+    permission_classes=(AuthCanViewEdit,)
+    serializer_class=serializers.ConceptSerializer
+    queryset=_concept.objects.all()
